+++ src/main/java/fuzs/deathfinder/DeathFinder.java
@@ -2,9 +2,7 @@ package fuzs.deathfinder;
 
 import fuzs.deathfinder.config.ClientConfig;
 import fuzs.deathfinder.config.ServerConfig;
-import fuzs.deathfinder.handler.DeathCompassHandler;
 import fuzs.deathfinder.handler.DeathMessageHandler;
-import fuzs.deathfinder.registry.ModRegistry;
 import fuzs.puzzleslib.config.ConfigHolder;
 import fuzs.puzzleslib.config.ConfigHolderImpl;
 import net.minecraftforge.common.MinecraftForge;
@@ -27,15 +25,11 @@ public class DeathFinder {
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
         ((ConfigHolderImpl<?, ?>) CONFIG).addConfigs(MOD_ID);
-        ModRegistry.touch();
         registerHandlers();
     }
 
     private static void registerHandlers() {
-        final DeathMessageHandler deathMessageHandler = new DeathMessageHandler();
-        MinecraftForge.EVENT_BUS.addListener(deathMessageHandler::onLivingDeath);
-        final DeathCompassHandler deathCompassHandler = new DeathCompassHandler();
-        MinecraftForge.EVENT_BUS.addListener(deathCompassHandler::onLivingDrops);
-        MinecraftForge.EVENT_BUS.addListener(deathCompassHandler::onPlayerClone);
+        final DeathMessageHandler handler = new DeathMessageHandler();
+        MinecraftForge.EVENT_BUS.addListener(handler::onLivingDeath);
     }
 }